feat: add wagmi adapter and confirmation blocks support (#17)
- Add wagmi adapter for public provider integration
- Implement confirmation blocks parameter for purchase functions
- Update purchase flow to support transaction confirmation monitoring
- Enhance Money class with improved arithmetic operations
- Update ERC20 ABI with proper type definitions
- Add x402 example project demonstrating express server integration
- Update version to 0.4.0-beta.0
- Update roadmap with completed features

+++ packages/client-sdk/src/adapters/wagmi-adapter/public-provider.ts
@@ -77,7 +77,7 @@ export class WagmiPublicProvider implements IPublicProvider {
           args: [address as `0x${string}`],
           chainId: networkId,
         });
-        return balance as bigint;
+        return balance;
       }
     } catch (error) {
       throw this._wrapError(error, 'getBalance', params);
